requirements.txt
# WhatsApp Bot con Twilio
twilio>=8.10.0
flask>=3.0.0

# Cliente HTTP para API de Playtomic
aiohttp>=3.9.0

# AI Chatbot
openai>=1.3.0

# Base de datos
sqlalchemy>=2.0.0
alembic>=1.12.0

# Tareas programadas
APScheduler>=3.10.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3
beautifulsoup4>=4.12.0
lxml>=4.9.0
